requirements.txt
altgraph==0.17.4
blinker==1.9.0
buildozer==1.5.0
click==8.1.8
cmake==3.31.6
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
distlib==0.3.8
dlib==19.24.6
face-recognition==1.3.0
face_recognition_models==0.3.0
filelock==3.16.1
Flask==3.1.0
fonttools==4.50.0
gunicorn==23.0.0
imageio==2.37.0
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.5
kiwisolver==1.4.5
lxml==5.2.2
MarkupSafe==3.0.2
matplotlib==3.8.3
MouseInfo==0.1.3
moviepy==2.1.2
msvc_runtime==14.40.33807
numpy==1.26.4
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==24.0
pefile==2023.2.7
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.3.6
proglog==0.1.10
psutil==6.1.1
ptyprocess==0.7.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
PyMsgBox==1.0.9
pyparsing==3.1.2
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-vlc==3.0.20123
pytweening==1.2.0
pywin32==306
pywin32-ctypes==0.2.2
schedule==1.2.1
setuptools==75.2.0
sh==2.0.7
six==1.16.0
tkintervideo==0.0.2
tqdm==4.67.1
typing_extensions==4.12.2
virtualenv==20.26.5
Werkzeug==3.1.3
wheel==0.44.0
